Add IDurableEntityContext.DispatchAsync() directly to interface
In a similar vein to PR #1422, this PR aims to transfer an extension
method to a direct interface implementation. This allows the method to
be directly mockable via tooling like Moq.

Addresses #1419
