+++ src/WebJobs.Extensions.DurableTask/ContextImplementations/DurableEntityContext.cs
@@ -4,7 +4,9 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Reflection;
 using System.Runtime.ExceptionServices;
+using System.Threading.Tasks;
 using DurableTask.Core;
 using Microsoft.Azure.WebJobs.Host.Bindings;
 using Newtonsoft.Json;
@@ -415,10 +417,92 @@ namespace Microsoft.Azure.WebJobs.Extensions.DurableTask
             return instanceId;
         }
 
+        async Task IDurableEntityContext.DispatchAsync<T>(params object[] constructorParameters)
+        {
+            IDurableEntityContext context = (IDurableEntityContext)this;
+            MethodInfo method = FindMethodForContext<T>(context);
+
+            if (method == null)
+            {
+                if (string.Equals("delete", context.OperationName, StringComparison.InvariantCultureIgnoreCase))
+                {
+                    Entity.Current.DeleteState();
+                    return;
+                }
+                else
+                {
+                    throw new InvalidOperationException($"No operation named '{context.OperationName}' was found.");
+                }
+            }
+
+            // check that the number of arguments is zero or one
+            ParameterInfo[] parameters = method.GetParameters();
+            if (parameters.Length > 1)
+            {
+                throw new InvalidOperationException("Only a single argument can be used for operation input.");
+            }
+
+            object[] args;
+            if (parameters.Length == 1)
+            {
+                // determine the expected type of the operation input and deserialize
+                Type inputType = method.GetParameters()[0].ParameterType;
+                object input = context.GetInput(inputType);
+                args = new object[1] { input };
+            }
+            else
+            {
+                args = Array.Empty<object>();
+            }
+
+#if !FUNCTIONS_V1
+            T Constructor() => (T)context.FunctionBindingContext.CreateObjectInstance(typeof(T), constructorParameters);
+#else
+            T Constructor() => (T)Activator.CreateInstance(typeof(T), constructorParameters);
+#endif
+
+            var state = ((Extensions.DurableTask.DurableEntityContext)context).GetStateWithInjectedDependencies(Constructor);
+
+            object result = method.Invoke(state, args);
+
+            if (method.ReturnType != typeof(void))
+            {
+                if (result is Task task)
+                {
+                    await task;
+
+                    if (task.GetType().IsGenericType)
+                    {
+                        context.Return(task.GetType().GetProperty("Result").GetValue(task));
+                    }
+                }
+                else
+                {
+                    context.Return(result);
+                }
+            }
+        }
+
         void IDurableEntityContext.Return(object result)
         {
             this.ThrowIfInvalidAccess();
             this.CurrentOperationResponse.SetResult(result, this.messageDataConverter);
+        }
+
+        internal static MethodInfo FindMethodForContext<T>(IDurableEntityContext context)
+        {
+            var type = typeof(T);
+
+            var interfaces = type.GetInterfaces();
+            const BindingFlags bindingFlags = BindingFlags.IgnoreCase | BindingFlags.Public | BindingFlags.NonPublic | BindingFlags.Instance;
+
+            var method = type.GetMethod(context.OperationName, bindingFlags);
+            if (interfaces.Length == 0 || method != null)
+            {
+                return method;
+            }
+
+            return interfaces.Select(i => i.GetMethod(context.OperationName, bindingFlags)).FirstOrDefault(m => m != null);
         }
 
         private void ThrowIfInvalidAccess()
