+++ src/WebJobs.Extensions.DurableTask/ContextInterfaces/IDurableEntityContext.cs
@@ -2,6 +2,8 @@
 // Licensed under the MIT License. See LICENSE in the project root for license information.
 
 using System;
+using System.Reflection;
+using System.Threading.Tasks;
 using Microsoft.Azure.WebJobs.Host.Bindings;
 
 namespace Microsoft.Azure.WebJobs.Extensions.DurableTask
@@ -169,5 +171,25 @@ namespace Microsoft.Azure.WebJobs.Extensions.DurableTask
         /// </exception>
         /// <returns>The instance id of the new orchestration.</returns>
         string StartNewOrchestration(string functionName, object input, string instanceId = null);
+
+        /// <summary>
+        /// Dynamically dispatches the incoming entity operation using reflection.
+        /// </summary>
+        /// <typeparam name="T">The class to use for entity instances.</typeparam>
+        /// <returns>A task that completes when the dispatched operation has finished.</returns>
+        /// <exception cref="AmbiguousMatchException">If there is more than one method with the given operation name.</exception>
+        /// <exception cref="MissingMethodException">If there is no method with the given operation name.</exception>
+        /// <exception cref="InvalidOperationException">If the method has more than one argument.</exception>
+        /// <remarks>
+        /// If the entity's state is null, an object of type <typeparamref name="T"/> is created first. Then, reflection
+        /// is used to try to find a matching method. This match is based on the method name
+        /// (which is the operation name) and the argument list (which is the operation content, deserialized into
+        /// an object array).
+        /// </remarks>
+        /// <param name="constructorParameters">Parameters to feed to the entity constructor. Should be primarily used for
+        /// output bindings. Parameters must match the order in the constructor after ignoring parameters populated on
+        /// constructor via dependency injection.</param>
+        Task DispatchAsync<T>(params object[] constructorParameters)
+            where T : class;
     }
 }
